add script to make a dynamic Makefile for building HTML files
The rule "HTML" on the Makefile would convert all markdown files
to HTML files every time it is run. This is bad because it would
convert markdown files that were not changed, thus, wasting time
converting these files again. The rule "watch" will watch for any
changes in any markdown files, then generate the HTML files if a
change occurs. This is terrible because it will take a long time
to do so even if only one file change (because it will convert all
the other files that haven't changed).

To handle this, I created a script to make a dynamic Makefile, called
Makefile2. The Makefile2 will a rule for every HTML being generated.
Thus, it will generate only the HTML files that have changed. Although
it may seen a little complex, it was not hard to make the script, and
it is a good solution since we do not know which and how many files
there will be in the src folder.